Made a visual representation of Unit turns taking place using GameTick logic
Migrated Unit System logic to Unit Component. Made Unit Sprite and Unit Stat components data onlt (will likely add set/get methods to stat values though).
On initialization, the GameTick system connects signals of subsystems so when signals are emitted by subsystems or the GameTick system, all relevant calls are completed. Even though the option to associate signals is available in the editor, doing so via code is a good way for us to single place to keep track of the majority of signal connects.
GameTick operates by emitting signals related to the phase that is started/finished, including the gametick itself. Units currently have one signal to process, started_gt_phase, and is associated with incrementing its tick_counter with its current speed value. This method is no longer necessary for MapGrid (as Godot hierarchy and tree work makes more sense to keep this and readied_units to the Units object rather than the MapGrid)