Created Initial Battle Menu logic
Moved processing of unit turn from GameTick System to UnitsComponent (should probably reconsider the name - Unit Manager might be better, might want to make another component instead in the hierarchy that does this processing though, a Units node is generally useful). With this change added Battle Menu basics with default Godot UI and buttons. Only Wait currently works but the signals for the others can be connected and used for other things.